:lipstick: Design: MainPage Design
